requirements.txt
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
Pillow==8.2.0
numpy==1.20.3
plotly==4.14.3

